Add paste with restore on macOS
The macOS implementations behind the shared TextInserter: the pasteboard
through the Swift helper on its own thread (read only while macOS allows
it without asking, nspasteboard.org markers, current host only, file
promises left out), CGEvent keystrokes with the layout's key of V and
typed text in 20-unit chunks, the focused window through the
Accessibility API, and the new SecureInputOn outcome at the final gate.
The helper's ABI goes to 3. AddTextInsertion() now runs on macOS too;
nothing there calls the inserter until add-macos-dictation.

+++ src/Pisum.Transcribe/TextInsertion/IClipboardService.cs
@@ -1,20 +1,24 @@
 namespace Pisum.Transcribe.TextInsertion;
 
 /// <summary>
-/// The Windows clipboard. The <c>Try*Async</c> methods return <see langword="false"/> or <see langword="null"/> when
-/// another process keeps the clipboard open for about 1 second.
+/// The system clipboard: the Windows clipboard, or the general pasteboard on macOS. The <c>Try*Async</c> methods
+/// return <see langword="false"/> or <see langword="null"/> when the clipboard can't be used right now, and the
+/// implementation logs why: on Windows, another process keeps the clipboard open for about 1 second; on macOS, the
+/// helper library isn't available, or, for a snapshot only, the system doesn't let the application read the pasteboard
+/// without asking the user.
 /// </summary>
 internal interface IClipboardService
 {
     /// <summary>
-    /// The clipboard sequence number, which Windows increments on every clipboard change. Readable from any thread.
+    /// A number that changes on every clipboard change: the clipboard sequence number on Windows, the pasteboard's
+    /// change count on macOS. Readable from any thread.
     /// </summary>
-    uint SequenceNumber { get; }
+    long SequenceNumber { get; }
 
     /// <summary>
     /// Copies the current clipboard contents, best-effort for formats other than text.
     /// </summary>
-    /// <returns>The snapshot, or <see langword="null"/> if the clipboard was busy.</returns>
+    /// <returns>The snapshot, or <see langword="null"/> if the clipboard could not be read.</returns>
     Task<ClipboardSnapshot?> TrySnapshotAsync();
 
     /// <summary>
@@ -22,7 +26,8 @@ internal interface IClipboardService
     /// </summary>
     /// <param name="text">The text.</param>
     /// <param name="excludeFromHistory">
-    /// Whether to keep the text out of clipboard monitoring, clipboard history and the cloud clipboard.
+    /// Whether to keep the text out of clipboard monitoring, clipboard history and the cloud clipboard, or on macOS out
+    /// of clipboard managers and Universal Clipboard.
     /// </param>
     /// <returns><see langword="true"/> if the text is on the clipboard.</returns>
     Task<bool> TrySetTextAsync(string text, bool excludeFromHistory);
